make it possible to create a non-sysroot toolchain
by setting WITH_SYSROOT=0 in config.sh.

unaware of any advantages of the --sysroot switch to a crosscompiler,
i'm at least aware of one big disavantage: if sysroot was used,
the full path to libstdc++.so will end up in .la files;
and if such a .la file is referenced from a package built with
--with-sysroot=/PATH/TO/ROOTFS with the intention to make libtool
search there for libs, the absolute libstdc++.la filename will
end up with /PATH/TO/ROOTFS/PATH/TO/MUSL-CROSS/lib/libstdc++.la -
which is of course not existing.